squire-engine/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "squire-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "typer>=0.12.5",
    "uvicorn>=0.30.6",
]

[project.scripts]
squire = "squire.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/squire"]
